Add pnpm-lock.yaml with project dependencies
The pnpm-lock.yaml file was created to maintain the integrity and resolve issues with the project dependencies. These include several importers and packages such as ethers, express, hardhat, and many more. This will ensure everyone working on the project will have the same package versions and dependencies, preventing "works on my machine" problems.